// app/webEF/interfaces.ts
// WARNING: This file was generated by a gulp task

import {DBSchema, DBContext, DBEntity} from './ef'
DBSchema.create('./webef/schema.json');

export interface Item {
	id?: number;
	description?: string;
	deadline: Date;
	done: boolean;
	userId?: number;
	user?: User;
	tasks?: Task[];
}

export interface ItemTable extends lf.schema.Table {
	id?: lf.schema.Column;
	description?: lf.schema.Column;
	deadline: lf.schema.Column;
	done: lf.schema.Column;
	userId?: lf.schema.Column;
}

export interface ItemContext {
	item?: ItemTable;
	user?: UserTable;
	task?: TaskTable;
}

export interface Task {
	id?: number;
	itemId?: number;
	description: string;
	done: boolean;
}

export interface TaskTable extends lf.schema.Table {
	id?: lf.schema.Column;
	itemId?: lf.schema.Column;
	description: lf.schema.Column;
	done: lf.schema.Column;
}

export interface TaskContext {
	task?: TaskTable;
}

export interface User {
	id?: number;
	name: string;
}

export interface UserTable extends lf.schema.Table {
	id?: lf.schema.Column;
	name: lf.schema.Column;
}

export interface UserContext {
	user?: UserTable;
}

export interface DBMasterContext {
	item?: ItemTable;
	user?: UserTable;
	task?: TaskTable;
}

export class MyDBContext extends DBContext<DBMasterContext> {
	constructor(){super('db')}
	public item = this.DBEntity<Item, ItemContext>('item', ["user","task"]);
	public task = this.DBEntity<Task, TaskContext>('task', []);
	public user = this.DBEntity<User, UserContext>('user', []);
}